разделила код на отдельные модули
